Moved card number validation into its own JS file
